# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Инспектор ИИ — API
  version: 1.0.0
  description: |
    Сервис сверки проектной (ПД), рабочей (РД) и исполнительной (ИД) документации по матрице 132 параметров
    (кейс «Мосстройнадзор»). Система **не выносит вердикт**: она готовит кандидатов в нарушения с карточкой
    доказательств, решение принимает инспектор.

    Порядок работы (асинхронная pull-модель, ТЗ 1.4):
    1. `POST /auth/login` → токен; дальше заголовок `Authorization: Bearer <токен>`.
    2. `POST /objects` — объект; `POST /documents/upload` — файлы со стадией → `process_id`.
    3. `GET /processes/{processId}` — статус (PENDING → PARSING → READY); `GET /processes/{processId}/result` — ответ.
    4. `GET /protocols/{protocolId}/checks?kind=candidate` — кандидаты; `PUT …/decision` — решение инспектора.
    5. `POST /protocols/{protocolId}/finalize`; `GET /protocols/{protocolId}/export?format=pdf|docx|xml|json`.

    Учётные записи стенда: inspector / inspector, supervisor / supervisor, admin / admin, ml / ml.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /documents/upload:
    post:
      tags: [Объекты и документы]
      summary: Загрузка или дозагрузка файлов стадии → process_id
      description: |
        Форматы PDF, DOCX, XML; не больше 50 МБ на файл и 200 МБ на пакет. Повреждённые PDF и дубли (тот же SHA-256)
        отклоняются с пояснением, остальные принимаются. Если `start` не равен `false`, сразу запускается проверка.
        После финализации протокола дозагрузка запрещена (409).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [object_id, stage, files]
              properties:
                object_id: { type: string }
                stage: { type: string, enum: [PD, RD, ID] }
                section: { type: string, enum: [AUTO, KR, AR, OV, VK, EOM, SS, GP, PB, POS, OTHER], default: AUTO }
                start: { type: string, enum: ['true', 'false'], default: 'true' }
                files: { type: array, items: { type: string, format: binary } }
      responses:
        '202':
          description: Файлы приняты, проверка поставлена в очередь
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadResult' }
        '409': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
        '422':
          description: Ни один файл не принят
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadResult' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        code: { type: string, example: PROTOCOL_FINALIZED }
        message: { type: string }
        details: {}
        request_id: { type: string }
...
    UploadResult:
      type: object
      properties:
        process_id: { type: string, nullable: true }
        object_id: { type: string }
        accepted: { type: array, items: { type: object } }
        rejected:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              reason: { type: string }
...